rollback

+++ SERVER/app.py
@@ -1,8 +1,8 @@
+from flask import Flask, make_response, jsonify, request, session
 from models import db, Recipe, User
 from flask_restful import Api, Resource
 from flask_migrate import Migrate
-from flask import Flask, request, session
-#from flask_cors import CORS
+from flask_cors import CORS
 import os
 
 BASE_DIR = os.path.abspath(os.path.dirname(__file__))
@@ -19,7 +19,7 @@ app.secret_key = os.environ.get('SECRET_KEY')
 db.init_app(app)
 
 migrate = Migrate(app, db)
-#CORS(app, supports_credentials=True)
+CORS(app, supports_credentials=True)
 
 @app.route('/')
 def home():
@@ -27,25 +27,29 @@ def home():
 
 @app.route('/recipes', methods = ['GET', 'POST'])
 def all_recipes():
-    user_session_id = session.get('user_id')
-    user_session = User.query.filter(User.id == user_session_id).first()
     if request.method == 'GET':
         recipes = Recipe.query.all()
-        return [r.to_dict() for r in recipes], 200
+        return [r.to_dict(rules = ['-recipes']) for r in recipes], 200
 
     elif request.method == 'POST':
         json_data = request.get_json()
+        ######Getting signed in user
+        user_session_id = session.get('user_id')
+        user_session = User.query.filter(User.id == user_session_id).first()
+
         new_recipe = Recipe(
             name = json_data.get('name'), 
             ingredients = json_data.get('ingredients'),
-            steps = json_data.get('steps'),
+
             image = json_data.get('image'),
             user_id = user_session_id,
             user = user_session
+
+
         )
         db.session.add(new_recipe)
         db.session.commit()
-        print(new_recipe.to_dict())
+
         return new_recipe.to_dict(), 201
     
 @app.route('/recipes/<int:id>', methods = ['GET', 'PATCH','DELETE'])
@@ -57,16 +61,19 @@ def recipe_by_id(id):
     
     if request.method == 'GET':
         return recipe.to_dict(),200
-    
     elif request.method == 'PATCH':
         json_data = request.get_json()
         for field in json_data:
             setattr(recipe, field, json_data[field])
         db.session.add(recipe)
         db.session.commit()
+
         return recipe.to_dict(),200
     
     elif request.method == 'DELETE':
+        
+        if recipe is None: 
+            return {'error': 'recipe not found'}, 404
         db.session.delete(recipe)
         db.session.commit()
         return {}, 204
